[INTERNAL] core/Lib: do not register classes with unqualified names
Some tests define controls with class names like "TestControl" or
"MyControl". For such unqualified names, the runtime can't derive a
library name. This causes an assertion failure in the constructor of
`sap/ui/core/Lib`. To avoid those, such names are now detected and the
registration and event are skipped.

Change-Id: I564968e59cfd0513913a99b45986ddbe6c5096ce