[INTERNAL][FIX]sap.ui.fl remove wrong content in compVariant
In RTA the content needs to be save because of the undo button. This
content should only save to the correct variant even when a new variant
is created. So the content need to be revert from the old display
variant till the last save action.

Also need to add the modified flag to the undo in the SaveAs Command.

Change-Id: I295c6eab41fec773376437192bd68e71c8d35c72
SNOW: DINC0163356