[INTERNAL][FIX]sap.ui.fl remove wrong content in compVariant
In RTA the content needs to be save because of the undo button. This
content should only save to the correct variant even when a new variant
is created. So the content need to be revert from the old display
variant till the last save action.

Also need to add the modified flag to the undo in the SaveAs Command.

Also add change: 6209151

Change-Id: I295c6eab41fec773376437192bd68e71c8d35c72
SNOW: DINC0163356
(cherry picked from commit 36c0338c74cf228cf113c79c9bc88fce27ab0567)
CR-Id: 002075125800003434942024