[INTERNAL] mdc.Link: use resource bundle in designtime
This change fixes an issue in the mdc.Link when used inside the SmartLink. When the user activates RTA the resource bundle text fot the settings action could not be loaded for some reason. Adding the resource bundle directly into the designtime will fix this issue. Also implementing this into the ObjectIdentifier as the texts are currently different.

PS 2: update message
PS 3: don't override ObjectIdentifier text

SNOW: DINC0185757

Change-Id: Ia6695827253a6a63e1981f71e59a1f595d8379d9