[INTERNAL] mdc.Link directNavigation handling
This fixes an issue with the directNavigation handling of the mdc.Link that came up in the demokit samples.

Adjusting the smoke test.
Fixing an issue in the beforeNavigationCallback of the demokit delegate.
The Dialog that is created in that callback should not be saved to the intance of the delegate.

Change-Id: I181d1a17c0cca607b58da8305ea733416fceece2