[INTERNAL] Integration Cards: Add attribute omitSandbox to WebPage content
The default behavior is to put all sandbox restrictions on the iframe.
But in some cases this is not desired
and lifting each restriction individually is not an option.
Therefore the need to have such an attribute.

This attribute was explicitly requested
by stakeholders for migration of
old Component cards to the WebPage card.

Change-Id: Ifb3d09afa22035847ecd02f17c44740878dd802d