[FIX] sap.m.SinglePlanningCalendarGrid: update appointment styling
Issue: When an appointment is created with a custom color, the
background color is added with JS which overwrites the selected
background color. Also there should be no additional background applied
to appointments when hovered, in Fiori & Horizon Light and Dark themes.

Solution: Only add custom background color when blocker or appointment
isn't selected, and remove hover background color that does not adhere
to the Visual Design specs.

SNOW: DINC0100369
SNOW: DINC0114097
Change-Id: If7b9bed526cc2a8124a72189c4a86a943c3719c6