[FIX] sap.uxap.ObjectPageLayout: lazy loading with correct scroll pos
Problem: When the user navigates from one OPL to another and then comes
back, in some cases, the scroll position is wrong and the lazy loading
functionality is executed with it, firing subSectionEnteredViewPort
event with Sections, which should not be in the viewport (if the page
was scrolled at the correct place).

Solution: When _triggerVisibleSubSectionsEvents is called, OPL is making
sure that, if a Section/SubSection is selected, OPL is scrolled to it -
scrollToSection is called with 0 ms delay (immediately).
This is needed because sometimes the selected Section/SubSection may
have changed its position when reopening (comming back) to the OPL. In
this case - position is changed, dues to the above Sections, which are
now less in height (they have not entered viewport, data for them is not
requested, so they are empty now, but when previously the user had
navigated out of this OPL the Sections were not empty).

SNOW: DINC0177489
Change-Id: Ic6c00a95b3531516400d56cba187a3afbb867802