[FIX] ControllerExtension support for TypeScript
ControllerExtensions are an important feature in freestyle UI5 and in particular in Fiori Elements applications, e.g. developed in S/4. With S/4 switching more and more to development in TypeScript, which we support and push very much, it is a critical issue that ControllerExtensions are not properly supported in TypeScript.

This correction makes a dummy function available in the UI5 type definitions, which can be used to solve the problem. (No influence on runtime code.)
It also adds a private method (no API change on JS side) to the runtime, which throws an error explaining what might have gone wrong. This also creates no runtime risk, as the method did not exist at all before.

BCP: n/a
CR-Id: 002075125800002783572024
Change-Id: Iffbac965b7df0c697e61454f0ebdf63f84706f3a
(cherry picked from commit b7c8ba23137aba2bc5c97b26283db1fec9b063db)