[FEATURE] sap.ui.unified.Calendar: Month/Year/Year Range pickers visual improvements
In calendars with a secondaryCalendarType set, the button text in the month picker will sometimes be cut off. Now whenever there is a secondary calendar set, the month picker will only render buttons in 2 columns to properly fit the text.

With this change we are also aligning the heights for month, year, and year-range pickers to match the regular calendar height, and centering their contents.

Visual tests already cover changes in Calendar, so there is no need
of additional tests.

JIRA: BGSOFUIBALKAN-8064
Change-Id: Icb42323ada1acfdcd383e44be86126aab0dd6948