[INTERNAL] Table: Fix scrolling behavior in case of subpixel calculation in Chromium
In some cases depending on the display scaling and browser zoom on Chromium the scrollTop of the vertical scrollbar is expressed in subpixels.
In such cases the table does not detect when being scrolled to the end. To fix this we have to round up the scrollTop value before the check.

Change-Id: Ia0d56138e3b55c26821d1e4ada2c202c2e9d4792