[FIX] sap.f.DynamicPage: Correct header placement after expanded state change
Scenario: In a web application with a dynamic page containing a table or other control set to 100% height and a collapsible header, the header is initially collapsed  (expandedHeader = false) into the title area and the header contains enough content to move the dynamic page content out of the view port, so that a scrollbar needs to be displayed.

Problem: When the header is collapsed initially, the dynamic page is considering/moving the header in the title area in the onAfterRendering hook. Upon expanding, this results in the scrollbar not appearing, even if the content exceeds the visible area.

Solution: Ensure the header is placed explicitly within the content area when expanded. This recalculates the content height, including the header's height, and displays a scrollbar if the content exceeds the viewport height.

SNOW: CS20240007621392
Change-Id: If94a8b18994676f73a41a75a55ff791418e31c28