[FIX] sap.m.Carousel: fix page disappearing when scrolling
Customer wants this fix into 1.108 due to a significant user base still relying on it. 
Although the bug isn't blocking progress, it affects the visual quality of the launchpad,
making the system appear messy. Improving this gives a better overall impression.

The solution relies solely on the standard CSS "transition", which prevents us from incorporating our tests to manage it.

As this issue is fixed in 1.108, customer wants to be sure that the next versions (1.120) will have the fix as well.

SNOW: CS20240008231123
SNOW: CS20240007652224
Change-Id: I70072e3744c353e1948b41381ff3b31c29a535a4
CR-Id: 002075125800003087932024