[INTERNAL] sap.m.Breadcrumbs: Location adjustment
During a previous BLI, it was introduced a change to the default behavior of the sap.m.Breadcrumbs that was considered undesirable. The update transformed the current location element of the breadcrumb into a link. This change was considered incompatible because a link to the current page is not particularly meaningful for users. Additionally, the new behavior, which refreshed the page when clicking on the current location element, was seen as both confusing and unnecessary.

To address these issues, the logic has been revised so that the current location element now displays as plain text by default, as it always has. Links are only added when explicitly requested.

Jira: BGSOFUIPIRIN-6590
Change-Id: I4ecaee8e3f2cfd84eff28acc1c85255666bbda87
(cherry picked from commit d82c1ce4bb9ccf8df4d6612e0b240cee7a771de1)
CR-Id: 002075125800003115162024
SNOW: N/A