[FEATURE] Make sap/ui/test/nextUIUpdate publically available
The original module was placed under "sap/ui/qunit", which is an invisible package and not shown in the developer guide.

Different developer guide entries however already propagate the
usage of the `nextUIUpdate´ module for control testing purposes.

With this change we correct the visibility of the test utils modules
by moving them to the "sap/ui/test" package.

Fixes: https://github.com/SAP/openui5/issues/4074
Change-Id: Id7bfd87d15ed6921ba5ffda8857c1a636a200c38
SNOW: DINC0193814