[INTERNAL] OPA: Enable module after API for stopping apps
Problem: When trying to tear down the iFrame/UIComponent from the "after" hook of the QUnit modudle, the iFrame/UIComponent are not destroyed.

Solution: The problem comes from the fact that - when QUnits are not
running (in the "after" hook), the OPA logic for looping the queue is
not working, as the queue is empty. As this logis is not in place, there
is no one to call the "success" functions of the objects, created in the
tear down function of the OPA, which destroy the iFrame/UIComponent.
That's why, now, is such cases (when queue is empty), we directly call
the "success" functions and return a resolved promise.

JIRA: BGSOFUIPIRIN-6296
Change-Id: Icfa90ae961aceed6b6ea28201d100f761f76ef81