[FIX] sap.f.FlexibleColumnLayout: Resize issue with iframe content fixed
Problem:
When the user resized the column separator and releases the mouse
(i.e. stops resizing) while the mouse cursor is above an iframe,
the "mouseup" (or "touchend" for thouch devices) is not fired.

Solution:
Disable the pointer events for the columns content during
resize of the columns, to ensure that iframes nested
in that content are never a target for the mouse/touch
events during resize.

SNOW: CS20240007927591
Change-Id: I1bf01528b57fc97dc7c6f83caa31c1c0666f794b