[INTERNAL] Prepare OPA self-test testsuite for UI5 2.0
So far, all OPA self-tests have been executed twice, once with
QUnit 1.18.0 and once with QUnit 2.3.2. With this change, this
duplication is deprecated and will be removed in UI5 2.0. Then, all
the tests will only be executed with the then-default version of QUnit.

The special handling of the action/Press tests has been replaced with
explicit configuration for both of them. The action/Press1 test has been
deprecated (for the same reason as above).

Tests from the demokit samples are no longer executed twice as they
anyhow use the QUnit version that is hard-coded in their HTML pages.

Change-Id: Ib31456dbf728e200bb820cd23dc7e3c5c2f4379f