[FIX] sap.m.Table: Announce required state of columns with 'custom' label controls
Fixes an issue that was not addressed with I314475d6cf41eedbad72d6e160f44c10dc0c3018.

If the focus is on the header cell, the required state of a column is also announced if the column header is not an instance of sap.m.Label, e.g. sap.ui.mdc.table.ColumnHeaderLabel, but has a required property.

Change-Id: I34a6ba89248a3222d8a2d6696f5d4c6e6996e078
SNOW: DINC0115442