[INTERNAL]sap.ui.rta: AddIFrameDialog URL parameters for V4 Models
When using parameters for building URLs on the AddIFrameDialog, if the
model is OData V4, the parameters can be resolved based on e.g. locale.
This leads to a discrepancy between what the user sees on the parameter
values table and what is actually used on the URL.

This change adds the parameters to the URL removing their type, which
prevents the automatic resolution of the parameters and ensures that
the value on the URL is the same displayed on the table.

JIRA: TEAM42-351
Change-Id: Ib9ca64516770f4167ac540d7c6d7aaa7c42723bf