[FIX] sap.f.DynamicPage: Syncronization issue corrected
Background:
The width of the DynamicPage title area is smaller if the page
needs to show a scrollbar on the side of the title.
Therefore, the above width is updated whenever the scrollbar
of the DynamicPage is toggled.

Problem:
Endless flickering due to endless loop of the following
3 adjustments:
1) change in width of the DynamicPage title-area
2) change in height of the DynamicPage title-area
   (due to height change of its nested filter bar)
3) change in height of the DynamicPage content-area
   (due to height change of its nested table with auto rows count)

The problem appears when the browser shows a scrollbar
in the interval between adjustments 2) and 3) above,
which makes the page unnecessarily update its title width
to a smaller width (to make room for that scrollbar).

The scrollbar dissapears right after adjustment 3) above,
but the dynamicPage is notified for the need to update
its title-width again only in the next scheduled call of
notifications from the ResizeHandler (i.e. after the 200ms interval)
During that interval, the title width remains incorrect
and causes another size adjustment of the content inside
the title-area (the nested filter bar) => endless adjustement

Solution:
In the DynamicPage, re-check for the need of scroolbar
(and respectively change the width of the title) synchonously,
right after all height adjustments (including those inside
the content area) following the ResizeHandler notifications
completed.

Change-Id: If5294f54b7db67dc289d846ce1a7d31f99871547
SNOW: CS20240006981749