[FIX] BindingInfo: Ensure that named models are only parsed once
This change ensure that a the extraction of a model name from a binding path
is only done once per BindingInfo instance.

The issue is noticed with XML templating after the following commit:
152de8c224990ae2ac2ce53c8c2afb6b4ba8d2bc

However, the underlying problem is also present when reusing
sap.ui.base.BindingInfo instances.

A named model is extracted from a binding string at the first occurrence
of the character ">".
In expression bindings however the character ">" can also be used for
comparisons, e.g. ">=".
Reusing an already parsed sap.ui.base.BindingInfo, which still contains
such characters in its "path" has led to a wrong extraction of an additional
model name.

SNOW: DINC0142903
Change-Id: Ib4a98bbde46d0186834e3801111284f8e5dfdb4e