[INTERNAL]sap.ui.dt/rta: addHover/removeHover API for Selection service
The Selection service now provides an API to add/remove hover effect on
elements. This can be used to, e.g., highlight a control in the app when
it is hovered on the control tree in Fiori Tools, making it easier for
the user to identify which control it is.

JIRA: TEAM42-355
Change-Id: I94f47c63c07f98343433f8e28d9204933ad1d439