[FIX] sap.f.FlexibleColumnLayout: Cleanup move listeners on destroy
Problem:
Error upon click on the qunits page after the completion
of all tests due to failure to cleanup listeners to the
mousemove/touchmove events.

Solution:
The problem seems to reproduce in the setup of the qunit only
(because the test mocks mousemove but omits mouseup (unexpected
browser behavior).
However, since it is possible to omit mouseup due to preceding
error/ destroy during interactive resize etc., fixed by ensuring
that the listeners are detached on destroy.

Change-Id: I1702cc25f72391833739ac346e545e669c7426e1