[FIX] sap.f.DynamicPage: Fix table overlapping
Problem: Table is getting overlapped by the DynamicPage header because
scroll-padding-top of the wrapper is not updated, when the header is
pinned.
Additionally, no related to the Incident, but seen in a Sample - the
stickySubheaderProvider is not sticked when header is pinned and user
scrolls down.

Solution: Scroll-padding-top of the wrapper is not updated because, when
pinning the header, the stickyHeaderProvider does not go in the sticky
area. This is not becoming a problem until the sticky content (in this
case, IconTabBar) is not gettind rerendered (without rerendering of the
whole DynamicPage). Upon rerendering _oSubHeaderAfterRenderingDelegate
is executed, which adjusts the sticky content and moves it to the right
place. However, _updateTitlePositioning is not called and the
scroll-padding-top of the wrapper is not updated, leading to overlapping
of the content of the DynamicPage.

SNOW: DINC0165833
CR-Id: 002075125800003113592024
Change-Id: Id2cc2878bbac8b09ef16b4382a6bf90b74c3c0c7
(cherry picked from commit 423b46a352f16356480aa33e14827135d73eb907)