[FIX] sap.m.TimePicker: fix cursor position when there is mask and no entry
Problem: When there is mask set to the TimePicker, and there is no
value entered, sometimes on focus of the input field, the cursor
doesn't go to the beginning of the input field (and mask) because
of unwanted mouse movement during the click of the button.

Solution: If there is mask set and there is nothing entered in the
TimePicker's input, the cursor is forced to move at the beginning
of the input field in onmouseup event handler.

Test snippet: #389598

SNOW: CS20240007797996
CR-Id: 002075125800003043502024
Change-Id: I8ea37be9d3db2c4345ec6f77bd56bb3259f69e86
(cherry picked from commit 98c64b988c0994ae49adcf5c3c4f3e1fa556e0b8)