[FIX] We've now fixed the minimum height issue for the no data page in uploadSet control
When an empty UploadSet is embedded within a Page the controls adds a default height which increases the page size resembles additional empty spaces.
When embedded within a Dialog the same space gets added along with an additional scrollbar and the user would need to mandatorily scroll and then select the UploadButton to upload a new file, hampering usability.

The minimum height of 500px which was set from the control when the NoData IllustratedMessage was displayed has been removed.
Now the control dynamically calculates the height based on the container height and sets the respective values.

SNOW: DINC0182815
CR-Id: 002075125900003893362024
Change-Id: I88f20ba485808257092d8316d54ea333b79b1069
(cherry picked from commit 842d2c6fee0df064faf94f96631ad81a8c51ce19)