[INTERNAL] Icon Explorer: improved overview page behavior
Previously, the side panel could've been opened only when clicking on an
item from the Grid (the filtered icons). Or manually.

Now, when we land on the Overview page and there is an icon in the
URL query, we automatically open the side panel with the icon
information. We can also open/close it manually as well as when we click
on a Grid item.

JIRA: BGSOFUIPIRIN-6614
Change-Id: I38bb062c8dbbe61ed9b8bc9aeb13b033a2d86229