[INTERNAL] ValueHelp: attach events once they are needed
If a ValueHelp is cloned, the selection on it's table is handled by the basic instancetoo, not only by the cloned one. This leads to runtome errors or changes on the wrong control.

Now the events to the inner control (Table, FilterBar and so on) are only attached if the ValueHelp is connected to field.
The events of the Content are attached if the specific content is shown.
On clone, the events are temporary detached to prevent a coling of the event handlers.

Change-Id: Ibcd6ef91d43491f03fab29b1caef365c90083c19
SNOW: DINC0182824
SNOW: CS20240008302096
CR-Id: 002075125900003444672024