[INTERNAL] sap.m.OverflowToolbar: Implement onfocusfail logic
Background: There are some cases, when the focus is not getting restored
properly either on the overflow button or on a certain control of the
OTB.
For example: when the OFT Popover menu is opened, then a Dialog is
opened and closed --> the focus should be restored on the OFT overflow
button, but it completely disappears.
Another case: Button in the OFT is clicked and it disappears (visible
property set to "false"). The focus should go to the next button, if
any, or the overflow button. Currently, focus completely disappears.

Solution: Implement the onfocusfail handler, provided recently by Core
team, in order to move the focus on the correct element.

JIRA: CPOUI5FRAMEWORK-486
SNOW: N/A
CR-Id: 002075125800003114982024
Change-Id: Iad75bd064eaecb24183c52df8c8168cc66c6cf1d
(cherry picked from commit aa323d4276246bee2d9dd0028e339a89bb72e900)