[FIX] Table: Size of filter field in column menu
When the AdaptationFilterBar is first displayed in the personalization dialog and then in the column menu, filter fields may decrease their size step by step till they reach their correct size (looks like an animation).
This happens, because if the fields have a bigger width, for example in the dialog, and are then displayed again in a smaller UI, for example the column menu, the sizing of the MultiInput does not work well with the layout of the AdaptationFilterBar.

An old solution for the personalization dialog was to set the width of the filter fields to 100% with CSS. This solution is applicable in the column menu case as well, and is therefore moved to the AdaptationFilterBar to be avialable wherever it is used.

Change-Id: Ice9dafff95f7529bf9b2f2582ddf801266addc46
SNOW: DINC0177595