[INTERNAL][FIX] ColumnMenu: Close popover before opening at another position
Previously, when clicking on a new column while a column menu was open, the menu would flicker due to the order of execution: updating the content first, then repositioning. Now, if a popover is open, it is closed before performing updates and reopening it to ensure a smooth transition.

SNOW: DINC0030393
Change-Id: Id78de06db1e13a0d7ceb5b81a188ecfec9bed073