[FIX] rowmodes.Auto: Height calculation with margins
If the table has elements in the extension area above the grid (MessageStrip of DataStateIndicator, or content in the extension aggregation), and these elements have a top or bottom margin, then the Auto mode calculated the available space incorrectly. As a result, the table did overflow it's container when it shouldn't, which led to further UI issues (flickering) in more complex layouts involving a DynamicPage.

The reason for this is that the height of elements that contain elements with a margin may not be affected by these margins. The browser returns a height (for the parent) as if there were no margins. This can be influenced by changing the styles.

Change-Id: I160380ca4dce303cc6cf86736b3120c3099630a0
SNOW: CS20240007958439
CR-Id: 002075125800003157502024