[FEATURE] CellSelector: Allow selection with Row/RowOnly
So far the Cell Selection can only be used in combination with row selection with a selector in front of the row.

Now, to select cells in Row/RowOnly selection behaviors in the GridTable, SPACE does not select the row anymore. If you want to select a row, you need to use ENTER instead. SPACE will now only select the focused cell.

JIRA: CPOUIFTEAMB-2300
Change-Id: Ib34c4687558113077a5f35c0aedbeca58794d74b