[FIX] sap.m.Fiori20Adapter: Performance issue fixed
Problem:
Error in thrown on mobile devices when the adapter
iterates a huge (2000+ items) list in the search
for adaptable content.

Solution:
The adapter should be skipping list-based controls as
they do not contain adaptable content. Therefore:
1) Added sap.m.SelectList to the skip-list
[of list-based controls to be skipped from adaptation]
2) Also added sap.ui.comp.variants.VariantManagement
as it renders a list of variant items. It renders those
items inside a sap.m.SelectList, but the
sap.m.SelectList is not accessible to the Fiori20Adapter
for direct inspection (only its parent and its children),
because it is not defined in any aggregation [public or
private], but is instead defined as a private field
=> the Fiori20Adapter only sees (and inspects) the parent
sap.ui.comp.variants.VariantManagement before inspecting
the children
=> added the parent type to the list [of list-based
controls to be skipped from adaptation]

SNOW: CS20230006635584
Change-Id: I26c9e235e2abc91398bd6149d26d41a8d4fc3627
(cherry picked from commit c321314cb48c00ca2abbf1d3d7a9c3748567d0b9)
CR-Id: 002075125800003393182024