[FIX] sap.uxap.ObjectPageHeader: fix duplicated buttons
Problem: Action Buttons of ObjectPageHeader are duplicated both in the header and in the overflow menu (which is unnecessary shown).

Solution: The problem happens because of an internal property change of Button (accessKey) in onBeforeRendering hook. When a property of the ActionSheet (overflow menu) Buttons get updated, the "_change" event is fired. ObjectPageHeader reacts on this event and adapts the overflow button.
There is a logic of checking if there is at least one Button in
the ActionSheet with visible property set to "true". It is weird that
visible properties of all Buttons actually stays "true", unless the
header is resized/rerendered and there is not enough space for all the action Buttons.

However, in this case:
--> ActionSheet is created (lazily) because on a previous step overflow menu is needed and opened.
--> objectTitle of ObjectHeader is updated
--> onBeforeRendering of all controls is executed
--> new Buttons are created (cloned from the original action Buttons) for the ActionSheetMenu and added to it
--> ActionSheet is invalidated
--> onAfterRendering of ObjectPageHeader is executed
--> overflow button is correctly hidden
--> ActionSheet is rerendering
--> onBeforeRendering of its Buttons is executed
--> as those are newly created Buttons, rendered for the first time, they do not have "accessKey" property set (it is only internally set in onBeforeRendering
--> change of"accessKey" property triggers "_change" event
--> ObjectPageHeader reacts to it with _adaptOverflow function, which checks if there is at least one Button with visible=true and shows the overflow menu. Both Buttons have visible=true (as it is only updated to false, if calculated previously that there is not enough space for all actions in onAfterRendering of ObjectPageHeader or resize).This is not the case here, as only the ActionSheet and its Buttons were invalidated and rerendered.

Now, _adaptOverflow function of ObjectPageHeader does not react on any property change of ActionSheet Buttons, unless it is the visible
property, which is the only one that matters whether or not the overflow button should be shown.

SNOW: DINC0215070
Change-Id: I965dd9a2519e88e053ce71d07de412390189d431