[INTERNAL] SampleTester: Stabilize tests on Safari
In Safari, if an iframe is destroyed before the resources it contains are fully loaded, the incomplete loading of these resources can cause the browser to hang or become unresponsive.

The code below does not resolve this issue (it needs to be addressed by the respective components), but aims to make tests rather stable by delaying cleanup/destruction of the Page content, respectively trying to keep the iFrames live a bit longer.

SNOW: DINC0110146
CR-Id: 002075125900002823552024
Change-Id: I1a2d6a841639085bbee4af10e430977180a0b6bb
(cherry picked from commit b2290fb67c3698cff975beff2577a75008b152d4)