[FIX] Demo Kit: Change info fixed
When opening a change info page from a url, containing a concrete
version the release versions dialog was not showing any data,
because neo-app.json in these distributions doesn't contain any
routes information.
Now we read the neo-app.json from the main directory of the url
as a workaround, and show all the released versions regardless of where
it is opened.

SNOW: DINC0216943
Change-Id: I03dd946bd75315b842ba270321c02e18bd0c24a5