[FIX] sap.uxap.ObjectPageLayout: lazyloading with content size update
Problem: "subSectionEnteredViewPort" event is not fired for visible
SubSection. The problem happens when one of the SubSections (in this
case - the first one) has big height and after some time it gets resized
and becomes smaller. At the time lazyLoading is executed the SubSection
still has bigger height.

Solution: It seems OPL does not call doLazyLoading, when the content
size is updated. However, in some situations, like this one, when
content size is updated (first SubSection becomes smaller in height),
a new one (in this case - the third one) also gets in the viewport.
That's why now OPL executes lazy loading when content size is updated,
in cases a new SubSection gets into view.

SNOW: DINC0158319
Change-Id: I21bb98ad3a914102f07f5ff9828721e969664cc4