[INTERNAL][FIX]sap.ui.fl fix wrong content in compVariant also for VENDOR and CUSTOMER_BASE
As in change 6166827 the content needs to be save because of the undo button. This
content should only save to the correct variant even when a new variant
is created. So the content need to be revert from the old display
variant till the last save action. With this change it is also fix in visual editor for VENDOR and CUSTOMER_BASE layer.

Change-Id: I196e5712236ab1f5b4782e5abc4c8341db4749e4
SNOW: DINC0163356
(cherry picked from commit 6bd020a8b5849695623ce98a26c55ff41360ac62)
CR-Id: 002075125800003453252024