[FIX] sap.f.DynamicPage: Fix pin button visibility
Problem: When pin button should not be shown
(preserveHeaderStateOnScroll=true), if for some reason DynamicPageHeader
gets invalidated and rerendered, the pin button gets shown (while it
should not be).

Solution: DynamicPage determines whether or not the pin button should be
shown. However, it alters the sapUiHidden CSS class directly on the DOM
ref of the pin button. When DynamicPageHeader gets invalidated and
rerendered, the pin button is also rerendered and it does not have the
sapUiHidden class anymore. So now we toggle the CSS class for hiding the
pin button on an instance level (the instance of Button), instead on the
DOM ref.
Additionally, as now the pin button instance remembers whether or not it
should have the sapUiHidden class, after rerendering of the DynamicPage,
we not only hide the pin button if necessary (as it was by default
always shown), but also show it if necessary (because if in the above
case preserveHeaderStateOnScroll is switched to "false" the pin button
should be shown).

SNOW: DINC0229040
Change-Id: I820727170af5a651cf27dcb50b30be231f635042