[FIX] sap.m.MessageView: Fix failing qUNit test
Change the getTitleRef when property activeTitle is set to true for a MessageListItem.
Due to a change in the sap.m.Link control, the text of the link is now rendered in a span inside the "a" element.
The MessageView can no longer rely on the direct link content to know if the text has ellipsis or no.

Change-Id: I218e733babb67b89e1ae0b1a75aebff22f5fa45f