[INTERNAL] TableDelegate: Return all selected contexts with ResponsiveTableType
If sap.m.Table#getSelectedContexts is called without parameters, it returns only the selected contexts of items that are in the aggregation. Instead, we need to get all selected contexts, just like with the GridTable.

The delegate now gets all selected contexts from the inner table by passing bAll=true to sap.m.Table#getSelectedContexts.

The delegate keeps rememberSelections=true on the inner table in all cases. If the binding parameter $$clearSelectionOnFilter is true and the binding is filtered, it deselects the header context. In this event, the sap.m.Table clears the entire selection, regardless of the rememberSelections setting.

Change-Id: I262fd51d30943eed0325c2c5ebc598a5ce9ad9bd
Jira: CPOUIFTEAMB-2370