[FIX] sap.uxpa.ObjectPageSubSection: Improved key handling
Problem: When we are on a table cell with input element and press
'F7' key, instead of focusing the table row, we are focusing the
ObjectPageSubSection.

Root Cause: There is custom 'F7' handling in the ObjectPageSubSection.

Solution: We are executing the custom handling only on non marked events.

SNOW: DINC0201466
Change-Id: I83b7751366de0c2dca46cbf537876eca87b9539f