[FIX] sap.uxap.ObjectPageLayout: Fix wrong scrolling to Section
Problem: When users scrolls down and header gets snapped, then user
expands the header manually and press on the AnchorBar to scroll to a
lower Section, the scroll position is not correct (the user lands above
the target Section, seeing parts of the upper Section, instead on the
very top of the target Section).

Solution: scrollToSection method is called with the current (correct)
top position of the target Section. However, just before the scrolling
is executed, the OPL moves the stickied header back to the header
content place, which changes the actual scroll position of the target
Section.
That's why, now OPL recalculates the position of the target Section,
adding the header content part to it.

SNOW: DINC0229777
Change-Id: Ic9be8f758e20bc57c390538be7c9701deeb9a08a