[INTERNAL]Flex Support Tools: Fix dependency extraction
Due to recent refactoring in FL, the dependency extraction was broken.
It is still unclear which changes should be displayed on the
module, but to enable it to work again, the dependencies are
retrieved for the VM-independent changes + the changes from the
currently selected variant. This information will also be added to the
extension's UI.

In the future this can be improved by
returning the changes and the variant that they belong to, so we can
display them more accurately in the UI.

Fixes issue #90 in Flextension repository.

Change-Id: I154e0f19f16ea17cf2530fdfb2b79dcf90c98a15