[FIX] sap.m.Popover: Height calculations for scrollable contents are preserved after invalidation
The original change in master was a fix related to a customer issue about the size calculation of the Popover being incorrect, when it's  invalidated after opening.
The customer required the fix for 1.108. This change is a continuity downport to ensure consistency between versions from 1.108 above.

SNOW: CS20240008082446
Change-Id: I3f21ee5b27c6794ec4edb8722a4227b47d97c01d
CR-Id: 002075125900003018612024
(cherry picked from commit 5772485fc376eab306bdae1a0484c0c4be05ccf9)