[INTERNAL] mdc Chart: support DrillDown and DrillUp
This change will make use of the drilldown and drillup on the inner Chart.
You can now select some data points and open the View By list to make a drilldown to this Dimension with additional filters.
When you not select any data Point the behavior is as before and only the new  Dimension will be added as new visible Dimension.
By clicking on the breadcrumbs you can now also drillup on the inner Chart drillstack (when you before made a drilldown)

Change-Id: I152cb953140eac5fc4a730299b5f17c34bb12071