[INTERNAL][FIX]sap.ui.dt: ObjectPage Integration Test stability fix
During test execution the output of the assert statements modify the
test page DOM and could lead to changes in the position of the elements.
To ensure that the test is stable, the positions are now collected before
any assert statement to make sure that the values will not change.

SNOW: DINC0239557
Change-Id: I188e9f0f728e288b2d906e93af32ca7995b1bace