[FIX] sap.m.QuickView: Title and description wrap correctly
Problem: The title was cut off with no truncation on the second line. The description was also cut off.

Solution: Truncate the title and description according to the visual design. Belize - title and description truncate on 2 rows. Horizon and Quartz - title truncates on 3 rows, description on 2 rows.

SNOW: DINC0231243
SNOW: DINC0238875
Change-Id: Ia14af9d9a6e2dc82f05713280d1cfe3e279f6625