[INTERNAL][FIX]sap.ui.rta: Incorrectly selectable Overlays in CViz
When in Change Visualization Mode, it is possible to tab through the
Overlays which have no change indicators, which is not intended.
This fix deactivates the focusability of Overlays when CViz is started,
setting it only on Overlays with indicators. When CViz is stopped, the
previous focusability is restored.

SNOW: DINC0240109
Change-Id: I9056d1997aa77ac5716632952bae56a17af7800d