[FIX] sap.f.DynamicPage: Fix Popover scroll in sample
Problem: In DynamicPageFreeStyle sample, there is a Popover, which is
opened, when the GenericTag is clicked. It has unwanted horizontal
scrollbar.

Solution: The unwanted scrollbar appears due to the fixed with of 300px
set to the Card, placed in the Popover. As the Popover content is also
with fixed width of 300px and there is a 1px border of each side of the
Popover.
Now we use 100% width for the Card, so it can take the whole space of
the Popover, without showing an unwanted scrollbar.

SNOW: DINC0242759
Change-Id: I89e14ac60a96d02c641d1d2f6cd361a8337d7e1f