[FIX] We now ensure that UploadSet does not insert or removes any item if they are bound to a model
when the items are bound to the model we should only fire an event to inform the application that the "delete" button of an item is pressed. Only the application can decide what can be done to the item correctly.

SNOW: DINC0034917
CR-Id: 002075125900002722632024
Fixes: https://github.com/SAP/openui5/issues/3939
Change-Id: Ic568c01bf881dfe4e04521bfd684cd7fb42ea3dc
(cherry picked from commit f44e71d88fca4d613a77f8edd2842dbb6605df16)