[INTERNAL] Tables: Support for SAP Companion
By default, field helps are displayed at the control having property
bindings, or the control whose aggregation is bound (containing controls
with property bindings).
In tables, the field helps should be displayed at the column header.

If a column is moved to the popin, the column header cannot be used to
display the field help, so the fieldHelpDisplay association is cleared
to fall back to the default behavior.

Just like the ariaLabelledBy association, the fieldHelpDisplay
association is set only on the cell control itself, not on any ot its
children. It's the responsibility of the user to set it in this case if
necessary. However, in the demo app provided to test the companion, it
was sufficient to set the association on the cell only (e.g. VBox).
The association is not set or changed by the tables if it's set by the
user.

Change-Id: I76ea7c4324dc9e1d8ef68a4cd5315b2709398ab9
Jira: CPOUIFTEAMB-2358