[FIX] sap.m.ViewSettingsDialog: fixed "select all" checkbox positioning
Problem: There was a misplacement of "select all" checkbox located in
the Filter By section of ViewSettingsDialog caused by another change
related to toolbar.

Solution: CSS selectors are precized in ViewSettingsDialog CSS file
to achieve proper positioning.

Note: not critical but it is good to be fixed, pure CSS fix.

SNOW: DINC0245398
CR-Id: 002075125900002533392024
Change-Id: If7b133374562f8683e77d5279f6f9f349eec3d57