[INTERNAL][FIX]sap.ui.fl: Multiple VM controls share a single URL param
Until now there was a URL parameter for each VM controls, for example:
sap-ui-fl-control-variant-id=var1&sap-ui-fl-control-variant-id=var2

This caused a console error in FLP as parameters are not supposed to
be repeated. With this change, we will have a single URL parameter
with the values separated by commas, like:
sap-ui-fl-control-variant-id=var1,var2

Legacy URLs with multiple parameters will still be supported.

SNOW: DINC0240075
Change-Id: I879ae5860d84d9edbbd8ebed03cb2ad3f79fc00c